requirements.txt
python-binance
typer
rich